Move `spawn_zebrad_for_rpc_without_initial_peers`
Place it in the `launch` sub-module.
