+++ zebrad/tests/common/lightwalletd/send_transaction_test.rs
@@ -12,12 +12,9 @@
 //! already been seen in a block.
 
 use std::{
-    collections::HashSet,
     env,
-    net::SocketAddr,
     path::{Path, PathBuf},
     sync::Arc,
-    time::Duration,
 };
 
 use color_eyre::eyre::{eyre, Result};
@@ -30,23 +27,19 @@ use zebra_chain::{
     transaction::Transaction,
 };
 use zebra_state::HashOrHeight;
-use zebra_test::{args, command::NO_MATCHES_REGEX_ITER, prelude::*};
 
-use crate::{
-    common::{
-        cached_state::{
-            copy_state_directory, load_tip_height_from_state_directory,
-            start_state_service_with_cache_dir, ZEBRA_CACHED_STATE_DIR_VAR,
-        },
-        launch::ZebradTestDirExt,
-        lightwalletd::{
-            self, random_known_rpc_port_config,
-            rpc::{connect_to_lightwalletd, spawn_lightwalletd_with_rpc_server},
-            zebra_skip_lightwalletd_tests, LIGHTWALLETD_TEST_TIMEOUT,
-        },
-        sync::perform_full_sync_starting_from,
+use crate::common::{
+    cached_state::{
+        copy_state_directory, load_tip_height_from_state_directory,
+        start_state_service_with_cache_dir, ZEBRA_CACHED_STATE_DIR_VAR,
     },
-    PROCESS_FAILURE_MESSAGES, ZEBRA_FAILURE_MESSAGES,
+    launch::spawn_zebrad_for_rpc_without_initial_peers,
+    lightwalletd::{
+        self,
+        rpc::{connect_to_lightwalletd, spawn_lightwalletd_with_rpc_server},
+        zebra_skip_lightwalletd_tests, LIGHTWALLETD_TEST_TIMEOUT,
+    },
+    sync::perform_full_sync_starting_from,
 };
 
 /// The test entry point.
@@ -216,46 +209,6 @@ where
     };
 
     Ok(block.transactions.to_vec())
-}
-
-/// Spawns a zebrad instance to interact with lightwalletd, but without an internet connection.
-///
-/// This prevents it from downloading blocks. Instead, the `zebra_directory` parameter allows
-/// providing an initial state to the zebrad instance.
-pub fn spawn_zebrad_for_rpc_without_initial_peers<P: ZebradTestDirExt>(
-    network: Network,
-    zebra_directory: P,
-    timeout: Duration,
-) -> Result<(TestChild<P>, SocketAddr)> {
-    let mut config = random_known_rpc_port_config()
-        .expect("Failed to create a config file with a known RPC listener port");
-
-    config.state.ephemeral = false;
-    config.network.initial_mainnet_peers = HashSet::new();
-    config.network.initial_testnet_peers = HashSet::new();
-    config.network.network = network;
-    config.mempool.debug_enable_at_height = Some(0);
-
-    let mut zebrad = zebra_directory
-        .with_config(&mut config)?
-        .spawn_child(args!["start"])?
-        .bypass_test_capture(true)
-        .with_timeout(timeout)
-        .with_failure_regex_iter(
-            // TODO: replace with a function that returns the full list and correct return type
-            ZEBRA_FAILURE_MESSAGES
-                .iter()
-                .chain(PROCESS_FAILURE_MESSAGES)
-                .cloned(),
-            NO_MATCHES_REGEX_ITER.iter().cloned(),
-        );
-
-    let rpc_address = config.rpc.listen_addr.unwrap();
-
-    zebrad.expect_stdout_line_matches("activating mempool")?;
-    zebrad.expect_stdout_line_matches(&format!("Opened RPC endpoint at {}", rpc_address))?;
-
-    Ok((zebrad, rpc_address))
 }
 
 /// Prepare a request to send to lightwalletd that contains a transaction to be sent.
